+++ lib/mdns/src/behaviour/timer.rs
@@ -18,14 +18,11 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use std::{
-    marker::Unpin,
-    time::{Duration, Instant},
-};
+use std::time::{Duration, Instant};
 
 /// Simple wrapper for the different type of timers
 #[derive(Debug)]
-#[cfg(any(feature = "async-io", feature = "tokio"))]
+#[cfg(feature = "tokio")]
 pub struct Timer<T> {
     inner: T,
 }
@@ -45,13 +42,15 @@ pub trait Builder: Send + Unpin + 'static {
 
 #[cfg(feature = "tokio")]
 pub(crate) mod tokio {
-    use super::*;
-    use ::tokio::time::{self, Instant as TokioInstant, Interval, MissedTickBehavior};
-    use futures::Stream;
     use std::{
         pin::Pin,
         task::{Context, Poll},
     };
+
+    use ::tokio::time::{self, Instant as TokioInstant, Interval, MissedTickBehavior};
+    use futures::Stream;
+
+    use super::*;
 
     /// Tokio wrapper
     pub(crate) type TokioTimer = Timer<Interval>;
@@ -60,7 +59,7 @@ pub(crate) mod tokio {
             // Taken from: https://docs.rs/async-io/1.7.0/src/async_io/lib.rs.html#91
             let mut inner = time::interval_at(
                 TokioInstant::from_std(instant),
-                Duration::new(std::u64::MAX, 1_000_000_000 - 1),
+                Duration::new(u64::MAX, 1_000_000_000 - 1),
             );
             inner.set_missed_tick_behavior(MissedTickBehavior::Skip);
             Self { inner }
@@ -87,7 +86,7 @@ pub(crate) mod tokio {
         }
 
         fn size_hint(&self) -> (usize, Option<usize>) {
-            (std::usize::MAX, None)
+            (usize::MAX, None)
         }
     }
 }